made `src` directory and fixed pyproject.toml
