+++ README.md
@@ -5,7 +5,7 @@
 
 # event_model
 
-
+Data model used by the bluesky ecosystem.
 
 This is where you should write a short paragraph that describes what your module does,
 how it does it, and why people should use it.
