--- event_model/__main__.py
@@ -1,16 +1,0 @@
-from argparse import ArgumentParser
-
-from event_model import __version__
-
-__all__ = ["main"]
-
-
-def main(args=None):
-    parser = ArgumentParser()
-    parser.add_argument("-v", "--version", action="version", version=__version__)
-    args = parser.parse_args(args)
-
-
-# test with: python -m python3_pip_skeleton
-if __name__ == "__main__":
-    main()
