+++ src/event_model/__main__.py
@@ -1,23 +1,14 @@
-"""Interface for ``python -m event_model``."""
-
 from argparse import ArgumentParser
-from collections.abc import Sequence
 
 from . import __version__
 
 __all__ = ["main"]
 
 
-def main(args: Sequence[str] | None = None) -> None:
-    """Argument parser for the CLI."""
+def main(args=None):
     parser = ArgumentParser()
-    parser.add_argument(
-        "-v",
-        "--version",
-        action="version",
-        version=__version__,
-    )
-    parser.parse_args(args)
+    parser.add_argument("-v", "--version", action="version", version=__version__)
+    args = parser.parse_args(args)
 
 
 if __name__ == "__main__":
