+++ src/event_model/documents/datum.py
