+++ src/event_model/documents/event.py
