+++ src/event_model/documents/event_descriptor.py
