+++ src/event_model/documents/generate/typeddict_to_schema.py
