+++ src/event_model/documents/resource.py
