+++ src/event_model/documents/run_stop.py
