+++ src/event_model/documents/stream_datum.py
