+++ src/event_model/documents/stream_resource.py
