+++ src/event_model/schemas/bulk_datum.json
