+++ src/event_model/schemas/bulk_events.json
