+++ src/event_model/schemas/event.json
