+++ src/event_model/schemas/event_descriptor.json
