+++ src/event_model/schemas/resource.json
