+++ src/event_model/schemas/run_stop.json
