+++ src/event_model/schemas/stream_datum.json
