+++ src/event_model/schemas/stream_resource.json
