+++ src/event_model/tests/test_auth.py
