+++ src/event_model/tests/test_emit.py
