+++ src/event_model/tests/test_schema_generation.py
